requirements.txt
absl-py==2.3.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
anykeystore==0.2
apex==0.9.10.dev0
attrs==25.3.0
black==22.3.0
blobfile==3.0.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
colorama==0.4.6
contourpy==1.3.2
crcmod==1.7
cryptacular==1.6.2
cryptography==45.0.4
cycler==0.12.1
deepspeed==0.6.5
defusedxml==0.7.1
-e git+https://github.com/Pmobilee/HADM-Server.git@c3cb89260e9125940639198c860fa4a807fef982#egg=detectron2
einops==0.8.1
fairscale==0.4.13
fastapi==0.115.14
filelock==3.14.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
hjson==3.1.0
hupper==1.12.1
hydra-core==1.3.2
idna==3.10
iniconfig==2.1.0
iopath==0.1.9
Jinja2==3.1.6
jmespath==0.10.0
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
-e git+https://github.com/open-mmlab/mmcv.git@90d83c94cfb967ef162c449faf559616f31f28c2#egg=mmcv
mmcv-full==1.7.2
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.0
PasteDeploy==3.1.0
pathspec==0.12.1
pbkdf2==1.3
Pillow==9.1.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.2.3
pyramid==2.0.2
pyramid-mailer==0.15.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2023.4
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
repoze.sendmail==4.4.1
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rq==2.4.0
scipy==1.16.0
sentry-sdk==2.32.0
setproctitle==1.3.6
sh==2.2.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
timm==0.5.4
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.65.2
transaction==5.0
translationstring==1.4
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
velruse==1.1.1
venusian==3.1.1
wandb==0.20.1
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.1.3
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xformers==0.0.31
xops==0.0.1
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zope.deprecation==5.1
zope.interface==7.2
zope.sqlalchemy==3.1
